Add files via upload
The Snake and Ladder game is a classic board game that involves moving players' pieces across a board with numbered squares. The goal is to reach the last square first. The board has snakes and ladders, which move players backward and forward, respectively. This is completely developed by OOP's Concept using Java.